requirements.txt
# =============================================================================
# TrippixnBot - Dependencies
# =============================================================================
# Author: حَـــــنَّـــــا
# =============================================================================

# Discord
discord.py>=2.3.0

# API Server
aiohttp>=3.9.0

# OpenAI (for AI ping responder + embeddings)
openai

# Vector database for RAG
chromadb
